Add custom policy

+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -11,7 +13,8 @@ using System.Threading.Tasks;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-	IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+	IRestaurantsRepository restaurantsRepository, 
+	IRestaurantAuthorizationService authorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
 	public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
 	{
@@ -19,6 +22,9 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
 		var restaurant = await restaurantsRepository.FindByIdAsync( request.Id );
 		if (restaurant is null)
 			throw new RestaurantNotFoundException();
+
+		if (!authorizationService.Authorize(restaurant, ResourceOperation.Delete))
+			throw new ForbiddenException();
 
 		await restaurantsRepository.Delete(restaurant);
